feature(#166 Model Manipulation): two new function calls
# {"request_type": "provider_list", "value": "all"}
# {"request_type": "provider_details", "value": "all"}
